mctp-ncsi: Add --up argument, to bring interface up on init
A common use-case is to bring the interface up directly after creating
it, but this is difficult to script, as mctp-ncsi will usually be run in
the background; an `ip link set up` may race with the interface
appearing.

To help with this case, add a --up option, causing mctp-ncsi to bring
the interface up on creation.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>